requirements.txt
flask
python-cipclient
adb-shell
samsungtvws